perf(mobile): stop the active-workout screen's per-second tick from re-rendering every exercise card
Same class of bug as the program editor, worse here since it's continuous: the
screen's elapsed-time stopwatch (setState + setInterval, ticking every second)
lived at the top of ActiveWorkout, and the whole exercise/set list was rendered
inline in that same component — no memo boundary was even possible. Every
second, for the full workout duration, every exercise card's ~2
TextInputs-per-set re-rendered.

Extracted the per-exercise block into a new memoized ActiveExerciseCard
(mirroring ExerciseFormCard), and wrapped the two handlers that aren't raw
Zustand store actions (handleCompleteSet, the per-card layout callback) in
useStableCallback so the memo isn't defeated by a fresh closure every tick.

Verified on-device: a render-count probe showed 0 card re-renders across 4
ticks of the idle timer, and exactly 1 (the touched card only) when editing a
field — confirmed against the same 30-exercise/98-set program used to
diagnose the original editor bug. Same UI, same behavior, same eslint/tsc
posture — active.tsx net shrinks by ~120 lines.
